refactor(analysis): move content complexity unit tests
