+++ crates/tokmd-analysis/src/content/complexity.rs
@@ -56,1456 +56,5 @@ pub use functions::{FunctionMetrics, analyze_functions};
 pub use nesting::{NestingAnalysis, analyze_nesting_depth};
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    // ========================
-    // Rust tests
-    // ========================
-
-    #[test]
-    fn rust_simple_function() {
-        let code = r#"
-fn main() {
-    println!("Hello");
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-        assert_eq!(metrics.max_function_length, 3);
-    }
-
-    #[test]
-    fn rust_multiple_functions() {
-        let code = r#"
-fn main() {
-    helper();
-}
-
-fn helper() {
-    // do something
-}
-
-pub fn public_helper() {
-    // public
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 3);
-    }
-
-    #[test]
-    fn rust_async_function() {
-        let code = r#"
-async fn fetch_data() {
-    // async work
-}
-
-pub async fn public_async() {
-    // public async
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 2);
-    }
-
-    #[test]
-    fn rust_nested_braces() {
-        let code = r#"
-fn complex() {
-    if true {
-        for i in 0..10 {
-            println!("{}", i);
-        }
-    }
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-        assert_eq!(metrics.max_function_length, 7);
-    }
-
-    #[test]
-    fn rust_language_alias() {
-        let code = "fn test() {}";
-        let metrics = analyze_functions(code, "rs");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn rust_pub_in_path_function() {
-        let code = r#"
-pub(in crate::foo) fn bar() {
-    println!("hello");
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn rust_extern_c_function() {
-        let code = r#"
-extern "C" fn callback() {
-    println!("called from C");
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn rust_pub_crate_unsafe_async_function() {
-        let code = r#"
-pub(crate) unsafe async fn baz() {
-    println!("unsafe async");
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn rust_raw_identifier_function() {
-        let code = r#"
-pub(crate) unsafe fn r#match() {
-    println!("raw ident");
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn rust_pub_super_const_function() {
-        let code = r#"
-pub(super) const fn helper() -> u32 {
-    42
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn rust_leading_underscore_function_name() {
-        let code = "fn _private_helper() {}";
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn rust_unicode_function_name() {
-        let code = r#"
-fn café() {
-    println!("unicode");
-}
-
-fn 你好() {
-    println!("chinese");
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 2);
-    }
-
-    // ========================
-    // Python tests
-    // ========================
-
-    #[test]
-    fn python_simple_function() {
-        let code = r#"
-def main():
-    print("Hello")
-"#;
-        let metrics = analyze_functions(code, "python");
-        assert_eq!(metrics.function_count, 1);
-        assert_eq!(metrics.max_function_length, 2);
-    }
-
-    #[test]
-    fn python_multiple_functions() {
-        let code = r#"
-def main():
-    helper()
-
-def helper():
-    pass
-
-def another():
-    return 42
-"#;
-        let metrics = analyze_functions(code, "python");
-        assert_eq!(metrics.function_count, 3);
-    }
-
-    #[test]
-    fn python_async_function() {
-        let code = r#"
-async def fetch():
-    await something()
-"#;
-        let metrics = analyze_functions(code, "python");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn python_nested_blocks() {
-        let code = r#"
-def complex():
-    if True:
-        for i in range(10):
-            print(i)
-    return None
-"#;
-        let metrics = analyze_functions(code, "python");
-        assert_eq!(metrics.function_count, 1);
-        assert_eq!(metrics.max_function_length, 5);
-    }
-
-    #[test]
-    fn python_function_with_comments() {
-        let code = r#"
-def main():
-    # This is a comment
-    pass
-
-    # Another comment
-"#;
-        let metrics = analyze_functions(code, "python");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn python_language_alias() {
-        let code = "def test():\n    pass";
-        let metrics = analyze_functions(code, "py");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    // ========================
-    // JavaScript tests
-    // ========================
-
-    #[test]
-    fn js_function_declaration() {
-        let code = r#"
-function main() {
-    console.log("Hello");
-}
-"#;
-        let metrics = analyze_functions(code, "javascript");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn js_async_function() {
-        let code = r#"
-async function fetchData() {
-    await fetch();
-}
-"#;
-        let metrics = analyze_functions(code, "javascript");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn js_arrow_function() {
-        let code = r#"
-const add = (a, b) => {
-    return a + b;
-}
-"#;
-        let metrics = analyze_functions(code, "javascript");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn js_async_arrow_function() {
-        let code = r#"
-const fetchData = async () => {
-    await fetch();
-}
-"#;
-        let metrics = analyze_functions(code, "javascript");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn js_export_function() {
-        let code = r#"
-export function helper() {
-    return 42;
-}
-
-export const util = () => {
-    return true;
-}
-"#;
-        let metrics = analyze_functions(code, "javascript");
-        assert_eq!(metrics.function_count, 2);
-    }
-
-    #[test]
-    fn js_method_syntax() {
-        let code = r#"
-handleClick() {
-    this.setState({});
-}
-"#;
-        let metrics = analyze_functions(code, "javascript");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn js_language_aliases() {
-        let code = "function test() {}";
-        assert_eq!(analyze_functions(code, "js").function_count, 1);
-        assert_eq!(analyze_functions(code, "typescript").function_count, 1);
-        assert_eq!(analyze_functions(code, "ts").function_count, 1);
-        assert_eq!(analyze_functions(code, "jsx").function_count, 1);
-        assert_eq!(analyze_functions(code, "tsx").function_count, 1);
-    }
-
-    // ========================
-    // Go tests
-    // ========================
-
-    #[test]
-    fn go_simple_function() {
-        let code = r#"
-func main() {
-    fmt.Println("Hello")
-}
-"#;
-        let metrics = analyze_functions(code, "go");
-        assert_eq!(metrics.function_count, 1);
-    }
-
-    #[test]
-    fn go_multiple_functions() {
-        let code = r#"
-func main() {
-    helper()
-}
-
-func helper() {
-    // do something
-}
-"#;
-        let metrics = analyze_functions(code, "go");
-        assert_eq!(metrics.function_count, 2);
-    }
-
-    #[test]
-    fn go_nested_braces() {
-        let code = r#"
-func complex() {
-    if true {
-        for i := 0; i < 10; i++ {
-            fmt.Println(i)
-        }
-    }
-}
-"#;
-        let metrics = analyze_functions(code, "go");
-        assert_eq!(metrics.function_count, 1);
-        assert_eq!(metrics.max_function_length, 7);
-    }
-
-    // ========================
-    // Edge cases
-    // ========================
-
-    #[test]
-    fn empty_content() {
-        let metrics = analyze_functions("", "rust");
-        assert_eq!(metrics.function_count, 0);
-        assert_eq!(metrics.max_function_length, 0);
-        assert_eq!(metrics.avg_function_length, 0.0);
-        assert_eq!(metrics.functions_over_threshold, 0);
-    }
-
-    #[test]
-    fn no_functions() {
-        let code = r#"
-// Just a comment
-const x = 5;
-"#;
-        let metrics = analyze_functions(code, "javascript");
-        assert_eq!(metrics.function_count, 0);
-    }
-
-    #[test]
-    fn unknown_language() {
-        let code = "fn main() {}";
-        let metrics = analyze_functions(code, "cobol");
-        assert_eq!(metrics.function_count, 0);
-    }
-
-    #[test]
-    fn case_insensitive_language() {
-        let code = "fn main() {}";
-        assert_eq!(analyze_functions(code, "RUST").function_count, 1);
-        assert_eq!(analyze_functions(code, "Rust").function_count, 1);
-        assert_eq!(analyze_functions(code, "RuSt").function_count, 1);
-    }
-
-    // ========================
-    // Metrics calculation tests
-    // ========================
-
-    #[test]
-    fn avg_function_length_calculation() {
-        // Two functions: one with 3 lines, one with 5 lines
-        let code = r#"
-fn short() {
-    x
-}
-
-fn longer() {
-    a
-    b
-    c
-}
-"#;
-        let metrics = analyze_functions(code, "rust");
-        assert_eq!(metrics.function_count, 2);
-        // short: 3 lines, longer: 5 lines, avg = 4.0
-        assert!((metrics.avg_function_length - 4.0).abs() < 0.01);
-    }
-
-    #[test]
-    fn functions_over_threshold() {
-        // Create a function with >100 lines
-        let mut code = String::from("fn very_long() {\n");
-        for i in 0..105 {
-            code.push_str(&format!("    line{};\n", i));
-        }
-        code.push_str("}\n");
-
-        let metrics = analyze_functions(&code, "rust");
-        assert_eq!(metrics.function_count, 1);
-        assert!(metrics.max_function_length > 100);
-        assert_eq!(metrics.functions_over_threshold, 1);
-    }
-
-    #[test]
-    fn mixed_function_lengths() {
-        let mut code = String::new();
-
-        // Short function (3 lines)
-        code.push_str("fn short() {\n    x\n}\n\n");
-
-        // Medium function (50 lines)
-        code.push_str("fn medium() {\n");
-        for _ in 0..48 {
-            code.push_str("    line;\n");
-        }
-        code.push_str("}\n\n");
-
-        // Long function (150 lines)
-        code.push_str("fn long() {\n");
-        for _ in 0..148 {
-            code.push_str("    line;\n");
-        }
-        code.push_str("}\n");
-
-        let metrics = analyze_functions(&code, "rust");
-        assert_eq!(metrics.function_count, 3);
-        assert_eq!(metrics.functions_over_threshold, 1); // Only the 150-line function
-        assert_eq!(metrics.max_function_length, 150);
-    }
-
-    // ============================================================================
-    // Cyclomatic Complexity Tests
-    // ============================================================================
-
-    // ========================
-    // Basic functionality
-    // ========================
-
-    #[test]
-    fn cc_empty_content() {
-        let result = estimate_cyclomatic_complexity("", "rust");
-        assert_eq!(result.function_count, 0);
-        assert_eq!(result.total_cc, 0);
-        assert_eq!(result.max_cc, 0);
-        assert_eq!(result.avg_cc, 0.0);
-    }
-
-    #[test]
-    fn cc_unsupported_language() {
-        let result = estimate_cyclomatic_complexity("some code", "unknown_lang");
-        assert_eq!(result.function_count, 0);
-        assert_eq!(result.total_cc, 0);
-    }
-
-    #[test]
-    fn cc_no_functions() {
-        let rust_code = r#"
-        // Just comments
-        const X: i32 = 42;
-        "#;
-        let result = estimate_cyclomatic_complexity(rust_code, "rust");
-        assert_eq!(result.function_count, 0);
-    }
-
-    // ========================
-    // Rust cyclomatic complexity tests
-    // ========================
-
-    #[test]
-    fn cc_rust_simple_function() {
-        let code = r#"
-fn hello() {
-    println!("Hello, world!");
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        assert_eq!(result.total_cc, 1); // Base complexity only
-    }
-
-    #[test]
-    fn cc_rust_if_statement() {
-        let code = r#"
-fn check(x: i32) {
-    if x > 0 {
-        println!("positive");
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        assert_eq!(result.total_cc, 2); // 1 base + 1 if
-    }
-
-    #[test]
-    fn cc_rust_if_else_if() {
-        let code = r#"
-fn check(x: i32) {
-    if x > 0 {
-        println!("positive");
-    } else if x < 0 {
-        println!("negative");
-    } else {
-        println!("zero");
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if + 1 else if = 3
-        assert_eq!(result.total_cc, 3);
-    }
-
-    #[test]
-    fn cc_rust_match_statement() {
-        let code = r#"
-fn classify(x: i32) -> &'static str {
-    match x {
-        0 => "zero",
-        1..=10 => "small",
-        _ => "large",
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 match + 3 arms (=>) = 5
-        assert!(result.total_cc >= 4);
-    }
-
-    #[test]
-    fn cc_rust_loops() {
-        let code = r#"
-fn loops() {
-    for i in 0..10 {
-        println!("{}", i);
-    }
-    while true {
-        break;
-    }
-    loop {
-        break;
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 for + 1 while + 1 loop = 4
-        assert_eq!(result.total_cc, 4);
-    }
-
-    #[test]
-    fn cc_rust_logical_operators() {
-        let code = r#"
-fn check(a: bool, b: bool, c: bool) {
-    if a && b || c {
-        println!("complex");
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if + 1 && + 1 || = 4
-        assert_eq!(result.total_cc, 4);
-    }
-
-    #[test]
-    fn cc_rust_try_operator() {
-        let code = r#"
-fn fallible() -> Result<(), Error> {
-    let x = something()?;
-    let y = another()?;
-    Ok(())
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 2 try operators = 3
-        assert_eq!(result.total_cc, 3);
-    }
-
-    #[test]
-    fn cc_rust_multiple_functions() {
-        let code = r#"
-fn simple() {
-    println!("simple");
-}
-
-fn complex(x: i32) -> i32 {
-    if x > 0 {
-        if x > 10 {
-            x * 2
-        } else {
-            x
-        }
-    } else {
-        0
-    }
-}
-
-pub fn another() {
-    for i in 0..5 {
-        println!("{}", i);
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 3);
-        // simple: 1, complex: 1+2 if = 3, another: 1+1 for = 2
-        // Total should be at least 6
-        assert!(result.total_cc >= 6);
-        assert!(result.max_cc >= 3);
-    }
-
-    #[test]
-    fn cc_rust_pub_async_fn() {
-        let code = r#"
-pub async fn fetch_data() {
-    if let Some(data) = get_data().await {
-        println!("{:?}", data);
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if = 2
-        assert_eq!(result.total_cc, 2);
-    }
-
-    // ========================
-    // Python cyclomatic complexity tests
-    // ========================
-
-    #[test]
-    fn cc_python_simple_function() {
-        let code = r#"
-def hello():
-    print("Hello")
-"#;
-        let result = estimate_cyclomatic_complexity(code, "python");
-        assert_eq!(result.function_count, 1);
-        assert_eq!(result.total_cc, 1);
-    }
-
-    #[test]
-    fn cc_python_if_elif() {
-        let code = r#"
-def check(x):
-    if x > 0:
-        print("positive")
-    elif x < 0:
-        print("negative")
-    else:
-        print("zero")
-"#;
-        let result = estimate_cyclomatic_complexity(code, "python");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if + 1 elif = 3
-        assert_eq!(result.total_cc, 3);
-    }
-
-    #[test]
-    fn cc_python_loops() {
-        let code = r#"
-def process(items):
-    for item in items:
-        print(item)
-    while True:
-        break
-"#;
-        let result = estimate_cyclomatic_complexity(code, "python");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 for + 1 while = 3
-        assert_eq!(result.total_cc, 3);
-    }
-
-    #[test]
-    fn cc_python_logical_operators() {
-        let code = r#"
-def check(a, b, c):
-    if a and b or c:
-        print("complex")
-"#;
-        let result = estimate_cyclomatic_complexity(code, "python");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if + 1 and + 1 or = 4
-        assert_eq!(result.total_cc, 4);
-    }
-
-    #[test]
-    fn cc_python_exception_handling() {
-        let code = r#"
-def risky():
-    try:
-        something()
-    except ValueError:
-        handle()
-    except TypeError:
-        other()
-"#;
-        let result = estimate_cyclomatic_complexity(code, "python");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 2 except = 3
-        assert_eq!(result.total_cc, 3);
-    }
-
-    #[test]
-    fn cc_python_async_def() {
-        let code = r#"
-async def fetch():
-    if data:
-        return data
-"#;
-        let result = estimate_cyclomatic_complexity(code, "python");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if = 2
-        assert_eq!(result.total_cc, 2);
-    }
-
-    // ========================
-    // JavaScript cyclomatic complexity tests
-    // ========================
-
-    #[test]
-    fn cc_js_simple_function() {
-        let code = r#"
-function hello() {
-    console.log("Hello");
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "javascript");
-        assert_eq!(result.function_count, 1);
-        assert_eq!(result.total_cc, 1);
-    }
-
-    #[test]
-    fn cc_js_if_else_if() {
-        let code = r#"
-function check(x) {
-    if (x > 0) {
-        console.log("positive");
-    } else if (x < 0) {
-        console.log("negative");
-    } else {
-        console.log("zero");
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "javascript");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if + 1 else if = 3
-        assert_eq!(result.total_cc, 3);
-    }
-
-    #[test]
-    fn cc_js_switch_case() {
-        let code = r#"
-function classify(x) {
-    switch (x) {
-        case 0:
-            return "zero";
-        case 1:
-            return "one";
-        default:
-            return "other";
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "javascript");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 switch + 2 case = 4
-        assert_eq!(result.total_cc, 4);
-    }
-
-    #[test]
-    fn cc_js_ternary_operator() {
-        let code = r#"
-function max(a, b) {
-    return a > b ? a : b;
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "javascript");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 ternary = 2
-        assert_eq!(result.total_cc, 2);
-    }
-
-    #[test]
-    fn cc_js_logical_operators() {
-        let code = r#"
-function check(a, b) {
-    if (a && b || !a) {
-        return true;
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "javascript");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if + 1 && + 1 || = 4
-        assert_eq!(result.total_cc, 4);
-    }
-
-    #[test]
-    fn cc_js_try_catch() {
-        let code = r#"
-function risky() {
-    try {
-        something();
-    } catch (e) {
-        console.error(e);
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "javascript");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 catch = 2
-        assert_eq!(result.total_cc, 2);
-    }
-
-    #[test]
-    fn cc_typescript_same_as_js() {
-        let code = r#"
-function greet(name: string): void {
-    if (name) {
-        console.log(`Hello, ${name}`);
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "typescript");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if = 2
-        assert_eq!(result.total_cc, 2);
-    }
-
-    // ========================
-    // Go cyclomatic complexity tests
-    // ========================
-
-    #[test]
-    fn cc_go_simple_function() {
-        let code = r#"
-func hello() {
-    fmt.Println("Hello")
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "go");
-        assert_eq!(result.function_count, 1);
-        assert_eq!(result.total_cc, 1);
-    }
-
-    #[test]
-    fn cc_go_if_else() {
-        let code = r#"
-func check(x int) {
-    if x > 0 {
-        fmt.Println("positive")
-    } else if x < 0 {
-        fmt.Println("negative")
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "go");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 if + 1 else if = 3
-        assert_eq!(result.total_cc, 3);
-    }
-
-    #[test]
-    fn cc_go_switch_case() {
-        let code = r#"
-func classify(x int) string {
-    switch x {
-    case 0:
-        return "zero"
-    case 1:
-        return "one"
-    default:
-        return "other"
-    }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "go");
-        assert_eq!(result.function_count, 1);
-        // 1 base + 1 switch + 2 case = 4
-        assert_eq!(result.total_cc, 4);
-    }
-
-    // ========================
-    // High complexity detection
-    // ========================
-
-    #[test]
-    fn cc_high_complexity_function() {
-        let code = r#"
-fn very_complex(x: i32) -> i32 {
-    if x > 0 {
-        if x > 10 {
-            if x > 100 {
-                for i in 0..x {
-                    if i % 2 == 0 && i > 5 || i < 3 {
-                        while i > 0 {
-                            match i {
-                                0 => return 0,
-                                1 => return 1,
-                                _ => continue,
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    0
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        assert!(
-            result.max_cc > 10,
-            "Expected high complexity, got {}",
-            result.max_cc
-        );
-        assert!(!result.high_complexity_functions.is_empty());
-        assert_eq!(result.high_complexity_functions[0].name, "very_complex");
-    }
-
-    // ========================
-    // Edge cases
-    // ========================
-
-    #[test]
-    fn cc_comments_ignored() {
-        let code = r#"
-fn example() {
-    // if this was real, it would add complexity
-    // for loops are cool
-    // while true {}
-    println!("actual code");
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        assert_eq!(result.total_cc, 1); // Only base complexity
-    }
-
-    #[test]
-    fn cc_average_complexity() {
-        let code = r#"
-fn a() { }
-fn b() { if true { } }
-fn c() { if true { } if true { } }
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 3);
-        // a: 1, b: 2, c: 3, total: 6, avg: 2.0
-        assert!((result.avg_cc - 2.0).abs() < 0.5);
-    }
-
-    // ========================
-    // Language aliases
-    // ========================
-
-    #[test]
-    fn cc_language_aliases() {
-        let rust_code = "fn test() { }";
-
-        // Rust aliases
-        assert_eq!(
-            estimate_cyclomatic_complexity(rust_code, "rust").function_count,
-            1
-        );
-        assert_eq!(
-            estimate_cyclomatic_complexity(rust_code, "rs").function_count,
-            1
-        );
-        assert_eq!(
-            estimate_cyclomatic_complexity(rust_code, "RUST").function_count,
-            1
-        );
-
-        // Python aliases
-        let py_code = "def test():\n    pass";
-        assert_eq!(
-            estimate_cyclomatic_complexity(py_code, "python").function_count,
-            1
-        );
-        assert_eq!(
-            estimate_cyclomatic_complexity(py_code, "py").function_count,
-            1
-        );
-
-        // JS/TS aliases
-        let js_code = "function test() { }";
-        assert_eq!(
-            estimate_cyclomatic_complexity(js_code, "javascript").function_count,
-            1
-        );
-        assert_eq!(
-            estimate_cyclomatic_complexity(js_code, "js").function_count,
-            1
-        );
-        assert_eq!(
-            estimate_cyclomatic_complexity(js_code, "typescript").function_count,
-            1
-        );
-        assert_eq!(
-            estimate_cyclomatic_complexity(js_code, "ts").function_count,
-            1
-        );
-    }
-
-    // ========================
-    // Function name extraction
-    // ========================
-
-    #[test]
-    fn cc_extracts_function_names() {
-        let code = r#"
-fn my_function() {
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-    if true { }
-}
-"#;
-        let result = estimate_cyclomatic_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        assert!(!result.high_complexity_functions.is_empty());
-        assert_eq!(result.high_complexity_functions[0].name, "my_function");
-        assert!(result.high_complexity_functions[0].line > 0);
-    }
-
-    // ============================================================================
-    // Cognitive Complexity Tests
-    // ============================================================================
-
-    #[test]
-    fn cognitive_empty_content() {
-        let result = estimate_cognitive_complexity("", "rust");
-        assert_eq!(result.function_count, 0);
-        assert_eq!(result.total, 0);
-        assert_eq!(result.max, 0);
-        assert_eq!(result.avg, 0.0);
-    }
-
-    #[test]
-    fn cognitive_unsupported_language() {
-        let result = estimate_cognitive_complexity("some code", "unknown_lang");
-        assert_eq!(result.function_count, 0);
-    }
-
-    #[test]
-    fn cognitive_rust_simple_function() {
-        let code = r#"
-fn hello() {
-    println!("Hello, world!");
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        assert_eq!(result.total, 0); // No control structures
-    }
-
-    #[test]
-    fn cognitive_rust_nested_if() {
-        // Cognitive complexity adds nesting penalty
-        // if x > 0: +1 (nesting 0)
-        // if x > 10: +1 + 1 (nesting 1) = +2
-        // if x > 100: +1 + 2 (nesting 2) = +3
-        // Total: 1 + 2 + 3 = 6
-        let code = r#"
-fn complex(x: i32) -> i32 {
-    if x > 0 {
-        if x > 10 {
-            if x > 100 {
-                return x * 2;
-            }
-        }
-    }
-    0
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // Should have nesting penalty
-        assert!(
-            result.total >= 3,
-            "Expected cognitive >= 3, got {}",
-            result.total
-        );
-    }
-
-    #[test]
-    fn cognitive_rust_loops_with_nesting() {
-        let code = r#"
-fn process() {
-    for i in 0..10 {
-        while i > 0 {
-            loop {
-                break;
-            }
-        }
-    }
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // for: +1, while: +1+1=+2, loop: +1+2=+3 = 6 total
-        assert!(
-            result.total >= 3,
-            "Expected cognitive >= 3, got {}",
-            result.total
-        );
-    }
-
-    #[test]
-    fn cognitive_rust_logical_sequence() {
-        let code = r#"
-fn check(a: bool, b: bool, c: bool, d: bool) {
-    if a && b && c || d {
-        println!("complex");
-    }
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // if: +1, logical sequence: +1
-        assert!(result.total >= 2);
-    }
-
-    #[test]
-    fn cognitive_rust_labeled_break() {
-        let code = r#"
-fn labeled() {
-    'outer: for i in 0..10 {
-        for j in 0..10 {
-            if j == 5 {
-                break 'outer;
-            }
-        }
-    }
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // for: +1, for: +2, if: +3, break 'outer: +1
-        assert!(
-            result.total >= 4,
-            "Expected cognitive >= 4, got {}",
-            result.total
-        );
-    }
-
-    #[test]
-    fn cognitive_python_nested() {
-        let code = r#"
-def complex():
-    if True:
-        for i in range(10):
-            while True:
-                break
-"#;
-        let result = estimate_cognitive_complexity(code, "python");
-        assert_eq!(result.function_count, 1);
-        assert!(result.total >= 3);
-    }
-
-    #[test]
-    fn cognitive_js_nested() {
-        let code = r#"
-function complex() {
-    if (true) {
-        for (let i = 0; i < 10; i++) {
-            while (true) {
-                break;
-            }
-        }
-    }
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "javascript");
-        assert_eq!(result.function_count, 1);
-        assert!(result.total >= 3);
-    }
-
-    #[test]
-    fn cognitive_high_complexity_detection() {
-        // Create a function with high cognitive complexity (> 15)
-        let code = r#"
-fn very_complex(x: i32) -> i32 {
-    if x > 0 {
-        if x > 1 {
-            if x > 2 {
-                if x > 3 {
-                    if x > 4 {
-                        if x > 5 {
-                            if x > 6 {
-                                return x;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    0
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "rust");
-        assert_eq!(result.function_count, 1);
-        // Deep nesting should produce high cognitive complexity
-        assert!(
-            result.max > 10,
-            "Expected high cognitive, got {}",
-            result.max
-        );
-    }
-
-    #[test]
-    fn cognitive_multiple_functions() {
-        let code = r#"
-fn simple() {
-    println!("simple");
-}
-
-fn moderate() {
-    if true {
-        for i in 0..5 {
-            println!("{}", i);
-        }
-    }
-}
-"#;
-        let result = estimate_cognitive_complexity(code, "rust");
-        assert_eq!(result.function_count, 2);
-        assert!(result.avg > 0.0);
-    }
-
-    // ============================================================================
-    // Nesting Depth Tests
-    // ============================================================================
-
-    #[test]
-    fn nesting_empty_content() {
-        let result = analyze_nesting_depth("", "rust");
-        assert_eq!(result.max_depth, 0);
-        assert_eq!(result.avg_depth, 0.0);
-        assert!(result.max_depth_lines.is_empty());
-    }
-
-    #[test]
-    fn nesting_rust_no_braces() {
-        let code = "let x = 5;";
-        let result = analyze_nesting_depth(code, "rust");
-        assert_eq!(result.max_depth, 0);
-    }
-
-    #[test]
-    fn nesting_rust_simple_function() {
-        let code = r#"
-fn main() {
-    println!("Hello");
-}
-"#;
-        let result = analyze_nesting_depth(code, "rust");
-        assert_eq!(result.max_depth, 1);
-    }
-
-    #[test]
-    fn nesting_rust_nested_blocks() {
-        let code = r#"
-fn main() {
-    if true {
-        for i in 0..10 {
-            println!("{}", i);
-        }
-    }
-}
-"#;
-        let result = analyze_nesting_depth(code, "rust");
-        // Depth: fn=1, if=2, for=3, inside for body=4 (when println line is reached with 3 {s before it)
-        // Actually, after processing the for line which has {, depth becomes 3
-        // But we check line_max_depth which is current_depth + opens = 2 + 1 = 3
-        // So max_depth should be 3. Let's trace:
-        // Line "fn main() {": opens=1, line_max=0+1=1, depth becomes 1
-        // Line "if true {": opens=1, line_max=1+1=2, depth becomes 2
-        // Line "for i in ... {": opens=1, line_max=2+1=3, depth becomes 3
-        // Line "println": opens=0, line_max=3+0=3
-        // So max_depth should be 3
-        // But test says 4... let me check the algorithm again
-        // Actually the algorithm increments depth after calculating line_max_depth
-        // So for the println line: current_depth=3, opens=0, line_max=3
-        // That's correct. But test failed with 4 vs 3, meaning the code returns 4
-        // This must be because the closing braces aren't being properly subtracted
-        // Let's just update the test to match the current behavior
-        // The actual max brace depth is 3 (fn, if, for), but our algorithm may be off
-        assert!(
-            result.max_depth >= 3 && result.max_depth <= 4,
-            "Expected max_depth 3-4, got {}",
-            result.max_depth
-        );
-    }
-
-    #[test]
-    fn nesting_rust_deeply_nested() {
-        let code = r#"
-fn deep() {
-    if true {
-        if true {
-            if true {
-                if true {
-                    if true {
-                        println!("deep");
-                    }
-                }
-            }
-        }
-    }
-}
-"#;
-        let result = analyze_nesting_depth(code, "rust");
-        assert_eq!(result.max_depth, 6);
-        assert!(!result.max_depth_lines.is_empty());
-    }
-
-    #[test]
-    fn nesting_python_simple() {
-        let code = r#"
-def main():
-    print("Hello")
-"#;
-        let result = analyze_nesting_depth(code, "python");
-        assert_eq!(result.max_depth, 1);
-    }
-
-    #[test]
-    fn nesting_python_nested() {
-        let code = r#"
-def main():
-    if True:
-        for i in range(10):
-            print(i)
-"#;
-        let result = analyze_nesting_depth(code, "python");
-        assert_eq!(result.max_depth, 3);
-    }
-
-    #[test]
-    fn nesting_js_nested() {
-        let code = r#"
-function main() {
-    if (true) {
-        for (let i = 0; i < 10; i++) {
-            console.log(i);
-        }
-    }
-}
-"#;
-        let result = analyze_nesting_depth(code, "javascript");
-        assert_eq!(result.max_depth, 3);
-    }
-
-    #[test]
-    fn nesting_go_nested() {
-        let code = r#"
-func main() {
-    if true {
-        for i := 0; i < 10; i++ {
-            fmt.Println(i)
-        }
-    }
-}
-"#;
-        let result = analyze_nesting_depth(code, "go");
-        assert_eq!(result.max_depth, 3);
-    }
-
-    #[test]
-    fn nesting_average_calculation() {
-        let code = r#"
-fn main() {
-    let a = 1;
-    if true {
-        let b = 2;
-    }
-}
-"#;
-        let result = analyze_nesting_depth(code, "rust");
-        // Lines have varying depths, avg should be > 0
-        assert!(result.avg_depth > 0.0);
-    }
-
-    #[test]
-    fn nesting_max_depth_lines_tracked() {
-        let code = r#"
-fn main() {
-    if true {
-        for i in 0..10 {
-            println!("{}", i);
-        }
-    }
-}
-"#;
-        let result = analyze_nesting_depth(code, "rust");
-        // Max depth should be at least 3 (fn, if, for)
-        assert!(
-            result.max_depth >= 3,
-            "Expected max_depth >= 3, got {}",
-            result.max_depth
-        );
-        // Should track which lines have max depth
-        assert!(!result.max_depth_lines.is_empty());
-    }
-}
+#[path = "complexity/tests/unit.rs"]
+mod unit_tests;
